pyproject.toml
[project]
name = "ComfyUI"
version = "0.3.49"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.6.0",
    "addict>=2.4.0",
    "aiohttp>=3.11.8",
    "albucore>=0.0.23",
    "albumentations>=2.0.7",
    "av>=12.0.0",
    "blend-modes>=2.2.0",
    "chardet>=5.2.0",
    "color-matcher>=0.6.0",
    "colour-science>=0.4.6",
    "comfy-cli>=1.3.9",
    "comfyui-frontend-package==1.20.5",
    "comfyui-ib-customnodes>=0.2.1",
    "comfyui-tooling-nodes>=0.1.0",
    "comfyui-workflow-templates==0.1.18",
    "diffusers>=0.26.3",
    "dill>=0.4.0",
    "dlib>=19.24.8",
    "dynaconf>=3.2.11",
    "einops>=0.8.1",
    "face-alignment>=1.4.1",
    "filelock>=3.18.0",
    "flow-vis>=0.1",
    "ftfy>=6.3.1",
    "fvcore>=0.1.5.post20221221",
    "gitpython>=3.1.44",
    "huggingface-hub>=0.19.0",
    "imageio-ffmpeg>=0.5.1",
    "importlib-metadata>=8.6.1",
    "insightface==0.7.3",
    "ipywidgets~=8.1",
    "kornia>=0.7.1",
    "librosa>=0.11.0",
    "lmdb>=1.4.1",
    "loguru>=0.7.3",
    "matplotlib>=3.9.4",
    "matrix-client==0.4.0",
    "mediapipe>=0.8.0",
    "mss>=10.0.0",
    "nest-asyncio~=1.0,>=1.5.9",
    "networkx[default]~=3.0",
    "numba>=0.61.2",
    "numpy>=1.22.0",
    "omegaconf>=2.3.0",
    "onnx>=1.10.0",
    "onnx2torch>=1.5.15",
    "onnxruntime>=1.15.0",
    "onnxruntime-gpu>=1.15.0",
    "onnxruntime-gpu>=1.21.1 ; platform_machine == 'x86_64' and sys_platform != 'darwin'",
    "opencv-contrib-python>=4.11.0.86",
    "opencv-python>=4.7.0",
    "opencv-python-headless>=4.11.0.86",
    "pandas>=2.2.3",
    "piexif>=1.1.3",
    "pillow>=9.0.0",
    "pixeloe>=0.1.4",
    "psutil>=7.0.0",
    "pydantic~=2.0",
    "pygithub>=2.6.1",
    "pykalman>=0.10.1",
    "pymatting>=1.1.13",
    "python-dateutil>=2.9.0.post0",
    "pyyaml>=6.0.2",
    "rembg>=2.0.65",
    "requests>=2.25.0",
    "rich>=13.7.1",
    "safetensors>=0.3.0",
    "scikit-image>=0.24.0",
    "scikit-learn>=1.6.1",
    "scipy>=1.10.0",
    "segment-anything>=1.0",
    "sentencepiece>=0.2.0",
    "soundfile>=0.13.1",
    "spandrel>=0.4.1",
    "svglib>=1.5.1",
    "timm>=1.0.7",
    "tokenizers>=0.13.3",
    "toml>=0.10.2",
    "torch>=2.0.0",
    "torchaudio>=2.7.0",
    "torchdiffeq>=0.2.5",
    "torchsde>=0.2.6",
    "torchvision>=0.15.0",
    "tqdm>=4.65.0",
    "transformers>=4.35.0",
    "transparent-background>=1.2.4",
    "trimesh[easy]>=4.6.8",
    "typer>=0.15.3",
    "typing-extensions>=4.13.2",
    "tyro==0.8.5",
    "ultralytics>=8.3.121",
    "uv>=0.7.2",
    "wget>=3.2",
    "wrapt~=1.0",
    "yacs>=0.1.8",
    "yapf>=0.43.0",
    "yarl>=1.18.0",
    "comfy-script",
    "facexlib>=0.3.0",
    "xformers",
    "groundingdino-py @ file:///D:/codes/aivision/ComfyUI/.cache/groundingdino_py-0.4.0-py2.py3-none-any.whl",
]

[project.urls]
homepage = "https://www.comfy.org/"
repository = "https://github.com/comfyanonymous/ComfyUI"
documentation = "https://docs.comfy.org/"

[tool.ruff]
lint.select = [
  "N805", # invalid-first-argument-name-for-method
  "S307", # suspicious-eval-usage
  "S102", # exec
  "T",    # print-usage
  "W",
  # The "F" series in Ruff stands for "Pyflakes" rules, which catch various Python syntax errors and undefined names.
  # See all rules here: https://docs.astral.sh/ruff/rules/#pyflakes-f
  "F",
]
exclude = ["*.ipynb", "**/generated/*.pyi"]



[tool.uv.sources]
torch = [
  { index = "pytorch-cu126", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cu126", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchaudio = [
  { index = "pytorch-cu126", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
comfy-script = { path = "custom_nodes/ComfyScript", editable = true }

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true

